feat: implement useAppEvent hook and deprecate useFrame
- Added the `useAppEvent` hook for subscribing to PlayCanvas application events, including 'update', 'prerender', and 'postrender'.
- Updated documentation to reflect the deprecation of the `useFrame` hook, providing migration guidance for users.
- Introduced tests for the `useAppEvent` hook to ensure proper registration and error handling.
- Removed the `useFrame` hook and its associated tests.
